requirements.txt
qrcode[pil]==8.2
opencv-python==4.11.0.86
pyzbar==0.1.9
sentence-transformers==4.1.0
numpy==1.26.4
openai==1.82.0
tqdm==4.67.1
faiss-cpu==1.7.4
Pillow==11.2.1